tools/requirements.txt
requests==2.21.0
python3-indy==1.8.1
python3-wrapper-vcx==0.2.4
aiohttp==3.5.4
